Don't raise exception for twilio error 13224: number invalid.
We get this exception when talking to the Call Power backend, and
submitting an invalid US number.  We do in fact do some rudimentary
front-end checking of numbers to make sure the area code isn't 555 and
doesn't start with 0 or 1, but Call Power does more rigorous testing
which is why we were sometimes getting this exception.