Trigger click event to prevent intermittent failures
See https://github.com/teampoltergeist/poltergeist/issues/530. I think these tests are failing because of an interaction between the click and some kind of animation or css transition. Triggering the click event on the element solves the problem, though it may be less reliable for making sure the element is visible to users and actually clickable.