Wait for request to civicrm in mailing signup test
Feature tests were failing intermittently because the request to subscribe a new user in civi hadn't completed yet. There's no success event we can look for on the client so I added a wait loop.

Capybara removed wait_until, but the creator says it's okay to add it back in https://www.varvet.com/blog/why-wait_until-was-removed-from-capybara/