Allow developers to edit page slugs
https://github.com/EFForg/action-center-platform/issues/161
This change was made once before, but had to be reverted because it
caused some weird id bugs.  This change covers some extra cases, and
hopefully won't cause the same bugs; i was unable to replicate them
using what i understood of the steps, so, high hopes.

If the bugs reoccur, a next step might be to add a `custom_slug` field
to the ActionPage model, and implement
`should_generate_new_friendly_id?` as `!custom_slug?`.