Don't churn action page slugs
https://github.com/EFForg/action-center-platform/pull/364
I'm still not 100% on why a previous developer thought it necessary
to set `@actionPage.slug = nil` on every update, but I removed it and
everything seems to work ok.
Also added a check that title is present, because the form passes
every attribute, with empty strings if the attr hasn't changed.  This
caused the model to always think that `title_changed?` was true.