Hide state action behind feature flag
The Civic API is going to be shut down in a week or two. We need to remove
the corresponding code and decide on if/how to move forward with a new service
that provides representative info.

In meantime this turns off the state level actions from the user interface
(public and admin) and returns a 404 from the corresponding endpoint in the
tools controller when STATE_ACTIONS_ENABLED is unset.

In the future, we can swap out the backend logic and reenable by setting
the env var to true.